audio: fix timestamp calculation for sending
according to RFC 3551 the RTP timestamp is independent of
number of channels and encoding. fix the timestamp calculation
for outgoing RTP packets, by dividing the total number of samples
on number of channels.

the bug was not affecting calls with mono audio, but was very
likely to affect calls with stereo audio (e.g. OPUS)