h265: remove call to x265_cleanup, fix crash
unloading the h265.so module crashes on OpenBSD,
inside the libx265.* code.

this could be related to linking to both libavcodec
and libx265, where also libavcodec links to libx265.
the OpenBSD packages supply both static (.a) and
shared (.so) libraries, and perhaps there is a problem
here that the static version is picked during link time.

running baresip+h265.so with valgrind gives more than
1000 warnings so I doubt that it is free of memory leaks.

this commit is a work-around for #106 (fixes #106)

  https://github.com/alfredh/baresip/issues/106