ctrl_tcp: added new socket for sending events and DTMF
Split up the command processing and send events to
be able to send commands synchron and process the
asynchron events in another thread.
Adding a DTMF handler to be able to send the DTMF
events as well.
Because the sending of events is the same the function
sendEvent() created and used in both event handlers.
Extend function command_handler() to be able to parse
also shortcuts.