audio: remove audio_srate and audio_channels
remove these config options:

  audio_srate            8000-48000
  audio_channels         1-2

these options has no practical usage, except to limit the
enabled audio codecs to the range of samplerate/channels
as specified.

if a module has multiple audio codecs, and you only want
to use one of them, this can now be done with the
"audio_codecs" parameter to the "accounts" file.