contact: add support for current contact
This is an improvement to make the contacts UI a bit more
user-friendly.

It adds the concept of a "current contact", which can be
changed by pressing the keys < and >.

- Pressing 'D' will dial the current contact.
- Pressing 'M' will send message to the current contact.

Example output of listing all contacts (press 'C'):

--- Contacts: (3) ---
       Unknown  Alfred <sip:aeh@creytiv.com>
>      Unknown  Bob <sip:bob@builder.com>
       Unknown  Carl <sip:carl@supertruck.com>

the left-most character '>' indicates the current contact.

also the current contact is saved to disk, so that the
state is correct next time you start baresip.