Adding client ID, base topic, username and password to mqtt login options.
Base Topic is needed to be separate for every baresip instance if you want to connect multiple
instances of baresip to the same broker.
Client ID is needed to be separate since there can only be one client with a specific ID
at every time. If two baresip connects with the same client ID (and username) the second
connection will simply replace the first.