account,ua: add flag cuser_ua for unique contact user
A new accounts flag `cuser_ua` (by default true) can be used to disable
the UA pointer suffix for the contact user. The suffix makes the contact
user unique and thus the UA selection for incoming SIP requests stable
against chosing the wrong.

If `cuser_ua` is set to false, chosing the UA for incoming requests
has to be done by host and port comparision. This is less stable because all
UAs can have the same cuser and therefor the first one that matches will be
used. This setting fixes registration issues on a Siemens HIpath and
1und1.de (see https://github.com/baresip/baresip/discussions/3149)