ua: set cuser depending on accounts cuser_ua (commend)
Disable extension of UA pointer for cuser if account
parameter cuser_ua is set to 0.
Searching the correct UA for incoming calls has to be checked
by host and port instead of cuser if cuser_ua is set to 0,
because all UAs can have the same cuser and therefor the
first one from the list matching will be used.

Removing the issue on a Siemens HIpath.

Only for Commend.