conf: add pre proc switch for non-file config
If the new pre-processor switch EXTCONFIG is specified during compilation the
configuration is not read from the file system. Except the default
configuration is used which can be altered by means of the new re function
conf_set. With baresip function config_parse_conf the changes are applied.

Example code:

    libre_init();

    conf_configure();

    conf_set(conf_cur(), "sip_listen", "0.0.0.0:5060");
    conf_set(conf_cur(), "audio_channels", "1");
    conf_set(conf_cur(), "audio_srate", "8000");
    conf_set(conf_cur(), "module", "g711");
    conf_set(conf_cur(), "module", "alsa");
    conf_set(conf_cur(), "module", "menu");

    config_parse_conf(conf_config(), conf_cur());
    baresip_init(conf_config(), false);