Add cmake build system
Add enough of a cmake build to compile a useful library.  It is not
exhaustive and will require some modification for various systems
(notably, this is capable of building for linux and Android).  So long
as re and rem are sufficiently included, then it can produce a baresip
executable that can be run from the command line with no missing
modules.