reg: use sipreg_registered and sipreg_failed
The function reg_failed() has to use new libre function sipreg_failed() in
order to be compatible with Cisco REGISTER keep-alives. It returns true if a
REGISTER was not successful. It is not important if the REGISTER had the
expires field set positive or zero. For simplicity we use sipreg_registered()
now for reg_isok().