reg: copy the failed state to new instance of sipreg
After a (re-)REGISTER fails the Cisco fallback REGISTER keep-alives are started
if an fbregint was configured. But the failc counter of the new sipreg instance
should be greater than zero in order to let sipreg_failed() return true. Then
when the first Cisco fallback REGISTER succeeds, failc is reset to zero and
sipreg_failed() returns false. Then the upper baresip layer knows that the SIP
proxy is available again.