reg, ua: add UA events for Cisco fallback registrations
The Cisco fallback REGISTER keep-alives are able to detect, if a SIP proxy
becomes available again. Then the upper layer may use this information to
switch to this SIP proxy if its priority is higher than the current. E.g. to
switch back from fallback SIP proxy to the main SIP proxy. This commit adds
UA_EVENT_FALLBACK_OK for this purpose and UA_EVENT_FALLBACK_FAIL for
completeness.