serreg: switch back to accounts with prio lower than current
Suppose SIP proxy "A" has prio = a and SIP proxy "B" has prio = b and a < b.

Further suppose that "A" is not available, so we are registered to "B". When
"A" becomes available, we want to switch to proxy "A". This means we start
the periodic (re-)REGISTER to "A" and the Cisco REGISTER keep-alives to "B".