Serial registration (#1083)
* account: add prio for serial registration

A priority field is added to struct account. Default prio is 0.

* menu,account,ua: initial registration only for prio 0 accounts

* ua: add function ua_regfailed()

* reg: add account prio to registration info/warnings

* account: add fbregint to struct account

* reg: use sipreg_registered and sipreg_failed

The function reg_failed() has to use new libre function sipreg_failed() in
order to be compatible with Cisco REGISTER keep-alives. It returns true if a
REGISTER was not successful. It is not important if the REGISTER had the
expires field set positive or zero. For simplicity we use sipreg_registered()
now for reg_isok().

* reg: copy the failed state to new instance of sipreg

After a (re-)REGISTER fails the Cisco fallback REGISTER keep-alives are started
if an fbregint was configured. But the failc counter of the new sipreg instance
should be greater than zero in order to let sipreg_failed() return true. Then
when the first Cisco fallback REGISTER succeeds, failc is reset to zero and
sipreg_failed() returns false. Then the upper baresip layer knows that the SIP
proxy is available again.

* reg: set the fallback REGISTER interval configured in the account

* ua: add Cisco REGISTER keep-alives

This are periodic un-REGISTER requests (REGISTER with expires=0) in order to
test the availability of a SIP-proxy. When sometimes a re-REGISTER of a UA
fails, we are able to switch immediately to an available SIP proxy. The new
function ua_fallback() starts the periodic un-REGISTER requests.

* ua: start Cisco REGISTER keep-alives only if fbregint was set

* ua: start Cisco REGISTER keep-alives when transport changes

* account: start Cisco fallback registrations on startup

When module account reads the configuration at startup of baresip only the
registrations for accounts with priority zero should be startet. For other
accounts with a positive fbregint the Cisco REGISTER keep-alives should be
started.

* menu: command ua_new should start also Cisco REGISTER keep-alives

The user may specify a fbregint and prio parameters with command ua_new. If
both got positive values, the Cisco REGISTER keep-alives will be started for
this account.

* reg, ua: add UA events for Cisco fallback registrations

The Cisco fallback REGISTER keep-alives are able to detect, if a SIP proxy
becomes available again. Then the upper layer may use this information to
switch to this SIP proxy if its priority is higher than the current. E.g. to
switch back from fallback SIP proxy to the main SIP proxy. This commit adds
UA_EVENT_FALLBACK_OK for this purpose and UA_EVENT_FALLBACK_FAIL for
completeness.

* serreg: add serial registration module

* serreg: prevent endless loop on registration failure

* serreg: Cisco serial registration mode as fallback

This is a more intelligent serial registration mode that uses also the response
information of the periodic un-REGISTER requests (Cisco REGISTER keep-alives)
to decide which UA would be a possible fallback.

* serreg: switch back to accounts with prio lower than current

Suppose SIP proxy "A" has prio = a and SIP proxy "B" has prio = b and a < b.

Further suppose that "A" is not available, so we are registered to "B". When
"A" becomes available, we want to switch to proxy "A". This means we start
the periodic (re-)REGISTER to "A" and the Cisco REGISTER keep-alives to "B".

* docs: add module serreg to example config