Menu: corrections for ring tones and call status by means of a global call counter (#1102)
* ua: count calls globally for max_calls check

The maximum number of allowed calls can be set by the setting call_max_calls.
But the calls are counted in each user agent separately. E.g. If somebody wants
to disable parallel calls completey, he would set calls_max_calls to 1. But
anyway, parallel calls are possible through different user agents.

This patch changes the behavior to count the calls globally.

* menu: for an outgoing call only play ringback for the current call

If a re-dialing is triggered and in the meanwhile an incoming call was added,
the current tone (ringtone) should not be stopped and replaced by a ringback
tone. The redial_handler() also does not change the current UA but
UA_EVENT_CALL_INCOMING does which makes the difference.

* menu: remove redundant stop of play tone in cmd_hangup

Stopping of play tone is done in the ua_event_handler() exactly when
ua_hangup() successfully terminates the current call. Stopping in cmd_hangup()
is to much if the current UA has no call, but maybe other UAs.

* menu: on call termination resume adequate play tone

When the current call terminates the play tone has to be stopped. But if there
is another incoming or outgoing call left, the ringtone resp. the ringback tone
has to be resumed.

* menu: stop play tone also for call progress

The event UA_EVENT_CALL_PROGRESS should also stop the play tone like
UA_EVENT_CALL_RINGING does. But the ringback tone should not be started.

* menu: remove redundant call to menu_set_incall()

Use new function uag_call_count() to decide if we are in a call, also after
call termination. Also removes obsolete have_active_calls().

* menu: correct incall decision for update_callstatus()

* menu: do not start ringback tone during established call

If there is an established call or a call with early media and we start an
outgoing call, then do not start the ringback tone. It may bother the
conversation.