audio, stream: put to and get from jitter buffer in two threads
During a call there is the RTP receive thread (A) and the auplay thread (B).
Previously the two threads had been connected by an aubuf. Thread A put the
RTP packets in the jbuf and also did a get in the same thread. Thus jbuf did
not really work as jitter buffer. Thread A did also the decoding and pushed the
data into the aubuf. Thread B fetched the data from the aubuf and pushed it to
the audio device.

Now in libre the jbuf was made thread safe and it works now as adaptive jitter
buffer. What means that the latency is controlled in reaction on the measured
network jitter.

Now thread A pushes into the jbuf and thread B fetches the RTP packets from the
jbuf, decodes to audio frames and pushes these to the audio device.