Adaptive jbuf (#1112)
* stream: remove obsolete stream_jbuf_reset()

The function jbuf_flush() already flushes and resets the jbuf. And
changing min and max jbuf size is obsolete now.

* audio, stream: put to and get from jitter buffer in two threads

During a call there is the RTP receive thread (A) and the auplay thread (B).
Previously the two threads had been connected by an aubuf. Thread A put the
RTP packets in the jbuf and also did a get in the same thread. Thus jbuf did
not really work as jitter buffer. Thread A did also the decoding and pushed the
data into the aubuf. Thread B fetched the data from the aubuf and pushed it to
the audio device.

Now in libre the jbuf was made thread safe and it works now as adaptive jitter
buffer. What means that the latency is controlled in reaction on the measured
network jitter.

Now thread A pushes into the jbuf and thread B fetches the RTP packets from the
jbuf, decodes to audio frames and pushes these to the audio device.

* stream: add config wish size for jbuf

* audio, stream: add silence detection for adaptive jbuf

Adapting of the latency in the jbuf should be done during periods of silence.
This commit also adds a silence detection. Which is disabled by default because
it is the only O(n) complexity part, where n is the number of samples (sampc).

* stream, audio: call again jbuf_get() if it returns EAGAIN

If the jitter buffer wants to reduce stored RTP packets jbuf_get() returns
EAGAIN. Then baresip fetches and decodes a second packet. Later the silence
detection decides if a frame can be dropped to reduce the latency.

* audio: use a timer to decode RTP packets

In order to satisfy the real-time properties of the auplay thread, a timer is
started in the auplay_write_hanlder() with delay zero in order to decode the
RTP packets in the main polling thread.

* audio: add decoding thread with wait condition

A decoding thread is added that replaces the timer. The decoding thread waits
for a wake-up signal from the auplay thread with a wait condition.

* config: add setting jitter_buffer_type=<off|fixed|adaptive>

* stream, audio: make jbuf method selectable