play: support different sample rates for file_ausrc
The used ausrc which can be configured with file_ausrc may return the
sample rate of the opened audio file. The pointer to the ausrc_prm is used for
returning the sample rate and channel number.

Instead of starting auplay immediately, it is started as soon as the first
audio frame arrived. Then the sample rate of the file is known by the ausrc
and was returned already in the ausrc_prm. This allows to allocate the auplay
with the right sample rate. If the ausrc does not support this feature, the
configured sample rate is used for the auplay.