play: always read until the end of aubuf
The eof was set in the ausrc_error_handler(). And the tmr_polling() stops the
playback if it detects that eof was set. So it was very likely that the
aubuf was not processed completely. Now the eof is set in the
aubuf_write_handler() after the last frame was read.