Play aufile extended support (#1155)
* aufile: return srate, channels of wav file

* aufile: support also PCMA and PCMU

Conversion is done similar to aufile_load() in play.c.

* play: always read until the end of aubuf

The eof was set in the ausrc_error_handler(). And the tmr_polling() stops the
playback if it detects that eof was set. So it was very likely that the
aubuf was not processed completely. Now the eof is set in the
aubuf_write_handler() after the last frame was read.

* play: support different sample rates for file_ausrc

The used ausrc which can be configured with file_ausrc may return the
sample rate of the opened audio file. The pointer to the ausrc_prm is used for
returning the sample rate and channel number.

Instead of starting auplay immediately, it is started as soon as the first
audio frame arrived. Then the sample rate of the file is known by the ausrc
and was returned already in the ausrc_prm. This allows to allocate the auplay
with the right sample rate. If the ausrc does not support this feature, the
configured sample rate is used for the auplay.

* play: add config sample rate and channels for file_ausrc

A configuration setting for sample rate and channel number independent from
call audio is needed for audio sources which does not support returning of
sample rate and channel number (e.g. module gst).