Multicast module (#1231)
* multicast: command functions and multicast sender with audio source

Send RTP streams to user defined addesses.
Used codec is user defined too

Separation of a sender and audio source object will allow a exchangable
audiosource implementation.

* multicast: add priority receiving part of the multicat module

The receiveing part is splitted into a priority and receiving
logic (receiver.c) while the player is located in (player.c)

The user is able to change the priority directly while a multicast
is running. The module adopt the new priorities and will change
the running stream if necessaray.

Automatic listener can be configured in the config file.
The priorities depend on the order in the config file.