multicast: add call priority handling.
assign a priority to a call. in case of a high priority stream,
the module will set the DnD and noDial flag of uag. All active calls
will be set on-hold. At the end of a stream the module checks if
it can resume calls and disable DnD and noDial.