call,ua,menu: add CALL_ANSWERED event fixes ALSA issues
This commit adds a new call event for the session answer. This also fixes the
ALSA issue #1261.
There was a short period of time where the audio playback device was open twice,
first for the ringback tone and second for the calls audio playback. Now the
ringback tone is stopped earlier, with the CALL_ANSWERED event.