Menu fixes for play tones2 (#1294)
* menu: remove obsolete function menu_find_call(callid)

There is already a function uag_call_find(callid).

* menu: add utility function menu_find_call()

* menu: add active_call_test

* menu: fix start of ringback tone

The ringback tone should be started for UA_EVENT_CALL_RINGING only if:
- it is not played already and
- there is no established call that wasn't put on-hold locally.

* menu: stop any tone when there is a new incoming call

The current call tracked by menu.callid changes to point to the new incoming
call. Thus any tone started for a previous call should stop.

* menu: fix for start of call waiting tone

In case of SIP auto answer the function start_sip_autoanswer() decides also
if call waiting should be played or the auto answer beep or nothing if the
answer delay is very short.

* menu: remove ANSI bell alert

* config: remove menu_bell setting for ANSI bell alert

* menu: on call closed only update tones if current call terminated

Error tones should be played only if current call terminated. Then we check if
there is another call for which we should proceed to play a ring/ringback tone.

* menu, ua: use active call state to decide which tone to resume

Shortly the call-id of the active call in menu was added. Now we should play a
tone that fits to the call state of the active call.

* menu: simplify decision for ringtone or call waiting tone

Call waiting should be played only if there is currently an established call
that is not put locally on-hold.

* menu: doxygen fix for menu_find_call()

* menu: on call termination select other call

Fixes unstopped ringing if the peer hangups up the call to us during state
CALL_STATE_INCOMING.