Multicast call priority (#1291)
* ua: nodial flag to prevent outgoing calls

* ua: DnD (Do not Disturb) function implementation

* menu: new function to enable/disable DnD mode

* multicast: add call priority handling.

assign a priority to a call. in case of a high priority stream,
the module will set the DnD and noDial flag of uag. All active calls
will be set on-hold. At the end of a stream the module checks if
it can resume calls and disable DnD and noDial.

* config: add multicast call priority config parameter