Call event answered fixes alsa issue (#1299)
* Revert "call: start streams asynchronously (issue #1261) (#1267)"

This reverts commit 2f45951f8841b99be2a5084796788febf55b61b9.

The reverted commit was not enough to fix the overlap between ringback tone
playback and call playback. The ALSA issue #1261 persists. A workaround is to
configure ALSA to use dmix.

* call,ua,menu: add CALL_ANSWERED event fixes ALSA issues

This commit adds a new call event for the session answer. This also fixes the
ALSA issue #1261.
There was a short period of time where the audio playback device was open twice,
first for the ringback tone and second for the calls audio playback. Now the
ringback tone is stopped earlier, with the CALL_ANSWERED event.