fixes #296: introduce the ability to passthrough the avformat video
In this PR, we introduce the possibility to passthrough the video received by the avformat module without the need to decode and re-encode it.

To achieve this goal, three new settings were introduced:

* **pass_through**: when set to **yes**, video passthrough is enabled
* **rtsp_transport**: it can be set to **tcp**, **udp**, **udp_multicast**, **http** or **https** to force the kind of transport used to communicate with the camera (when RTSP is used)
* **profile_level_id**: to set the profile level of the output stream, for example **42002a** to set profile_idc to 42, profile_iop to 00 and h264_level_idc to 2a

The copyright of this work is shared by Media Magic Technologies <developer@mediamagictechnologies.com> and Divus GmbH <developer@divus.eu> and it's released under the same licence used by Baresip.