menu: return complete URI for commands dial,dialdir
This can be used to match with event CALL_LOCAL_SDP field `peeruri` to identify
the right call. Then the field `id` of this event then contains the correct
call-id.

If media nat is configured, then the call-id is not given, when these commands
return. The above process is a work-around to determine the call-id.