ua: select laddr with route to SDP offer address
For incoming calls (INVITE) we get an SDP offer and can use the contained IP
address `dst` to find a local address which has a route to `dst`.

The function `sdp_connection()` replaces the `sdp_af_hint()` and still
retrieves the address family from the SDP offer. As fallback use address family
of SIP message source.

Checking if the address family is supported can be done also in
`ua_call_alloc()`. This avoids similar code in `sipsess_conn_handler()`.