multicast,config: add separate jitter buffer configuration
If multiple baresip instances which are in near range receive a multicast RTP
stream, it should be possible to adjust their playback latency to achieve a
simultaneous audio playback on all devices.

On the other hand the jitter buffer configuration for a SIP call should be
optimized to minimum latency.