stream: RTP inbound telephone events should not lead to packet loss
Telephone-event packets also increment the RTP sequence number and thus have to
be put into the jitter buffer. Otherwise missing packets are detected wrongly.
This leads to audio PLC (Packet loss concealment) or double playback of an
audio frame.
If a telephone event packet is read from jitter buffer, it has to be ignored
by reading a second packet. So the `stream_recv_handler()` needs a return
value which can be EAGAIN.