menu: add an event for call transfer
This allows an application to match the involved calls in a call transfer.
call 1 is the call to the transfer initiator.
call 2 is the call to the transfer target.
The CALL_TRANSFER event that is currently available does not contain the call-
id of the target call. And it is not possible to add this call-id to the
TRANSFER event because module menu invokes the call to the target as reaction
on this event.