call: postpone alloc of audio/video if user@domain is dialed
Put allocation of audio/video and initialization of the streams in a function
call_streams_alloc(). If user@domain is dialed and a registrar-less UA is used,
then call_streams_alloc() is done in the sipsess_desc_handler(). In this case
this is the earliest point in time where the laddr can be chosen equal to the
source address of the SIP request.