set sdp laddr to SIP src address (#1645)
* call: replace pre-encoded SDP by sipsess_desc_handler() for INVITE

The sipsess_desc_handler() updates the laddr of the SDP session and encodes
the SDP just before the SIP INVITE is sent.

* ua: set laddr for incoming SIP OPTIONS

* call: sipsess_desc_handler() should not change STUN laddr

* ua: remove best effort laddr

- Like before, for registered UAs use the laddr of the registration.
- For outgoing registrar-less calls to user@domain now it is also possible
  that the address family and laddr are undefined during allocation of the
  call object. They will be set when the SDP is request by the SIP session.

* call: postpone alloc of audio/video if user@domain is dialed

Put allocation of audio/video and initialization of the streams in a function
call_streams_alloc(). If user@domain is dialed and a registrar-less UA is used,
then call_streams_alloc() is done in the sipsess_desc_handler(). In this case
this is the earliest point in time where the laddr can be chosen equal to the
source address of the SIP request.

* account: remove obsolete parameter mediaaf

* call: disable "S D P - O f f e r" print