serreg: fix fallback accounts (#1660)
* Revert "serreg: fix server selection in case all server were unavailable"

This reverts commit 20ab46a3ae7722350c8798c5d6273342a509f126.

The reverted commit causes a lot of toggling between prio 0 and prio 1 for
cisco mode if there are two configured Cisco Callmanager accounts.
The commit 20ab46a3ae7722350c8798c5d6273342a509f126 will be handled differently.

* serreg: all UA_EVENT_REGISTER_FAIL should be handled

The functions check_registrations() anyway checks if all accounts with current
prio failed. This fixes the issue of reverted commit
20ab46a3ae7722350c8798c5d6273342a509f126 in a correct way.

* serreg: register_curprio() returns zero as soon as one register was sent

This function should return an error if no REGISTER was sent or could be sent
successfully.

* serreg: sprio should be 'Prev successful prio'

If there was no successful registration so far, then do not set it to current
prio.

* serreg: do not reset current prio for fallback accounts

The current prio always was reset to zero in function reset(). This only
allowed switching to main accounts (with prio 0) in function fallback_ok().
Instead reset current prio to zero should only be done as soon as there is
one prio 0 account with fbregint=0 or not set.