menu: Play ringback in play device
When placing a call I expect to hear the ringing sound in the headset
(play device) but not in the alert device.

This provides a fix similar to #934 (which was initially fixed in
99487e221d0f8af1994769c9b6df34fe6eccb580). However, this fix was
basically reverted in 3bd44f1c95f464af66820dda242b2a2de1e0fc03 when
the menu_play helper was introduced.