menu: exclude given call when searching for active call (#1721)
When a UA event occurs menu needs to search for other active calls to decide
which tone should be played.
E.g. on UA_EVENT_CALL_CLOSED menu is looking for another active call while the
closed call may still be in CALL_STATE_ESTABLISHED. In this case the closed
call should be excluded from the search.

This fixes resume of ringback tone for multiple parallel outgoing calls, when
the first peer declines the call.