stream: also for JBUF_ADAPTIVE decode immediately
This makes JBUF_ADAPTIVE completely obsolete what is the final goal
of the current PR. This is necessary because otherwise no decoding
would be done if `jitter_buffer_type   adaptive` is configured.

An alternative would be to keep all the JBUF_ADAPTIVE functions in
stream.c and audio.c for a deprecation period.