menu ringback for parallel call (#1827)
* menu: resume ringback also if call is in CALL_STATE_EARLY

For the parallel call use case the ringback stopped when second callee canceled
the call.

* menu: instead of restart ringback proceed playing for parallel call

If second callee canceled the call, the ringback stopped and restarted for the
first callee. Because the user does not care about the order of the callees
it is better to proceed playing the ringback tone.