call, menu: make selective early media RFC-3261 conform
Due to RFC-3261 the SDP in the 183 Session Progress has to be equal to the SDP
in the 200 Ok. The correct solution is to answer the call with same SDP and
send immediately a re-INVITE if needed. E.g. in case of early video send
re-INVITE in order to activate the audio stream.