audio: remove unused aubuf for decoding
The aubuf for decoding and the rx_filter_thread was introduced in order to
combine the benefits of an adaptive audio buffer with an AEC as audio filter.
If a sound daemon like pulseaudio is used, the AEC should be integrated into
this daemon. But if baresip directly connects to ALSA the aubuf between the
filters and the audio player has to be fixed and as short as possible.

We think that the rx_filter_thread with the decoding aubuf has not enough
benefits for the lots of extra code that was necessary. Instead one could use
a modified ALSA module which integrates an AEC solution, like we did. Then the
AEC runs in the auplay thread.

So this commit is a cleanup for unused code parts.