play: ring tone fixes if file_ausrc is set (#2050)
* aufile: use aubuf_read_auframe()

The time stamps are set now by aubuf automatically. They don't have to be set
after reading.

* play: use new aubuf API and don't start auplay too early

This fixes playback of tones if `file_ausrc` is set in baresip config.

* gst: disable wish size for aubuf

Fixes dropped beginning of a played audio file. For playing a audio files it is
not wanted that old frames are dropped on first read.

* play: move restart of ausrc to main thread

This fixes
- the warning: "main: thread check: called from a NON-RE thread without
  thread_enter()!"
- the ringing stops after seconds repeat
- underruns during file playback with ausrc

This fixes also underruns after the restart and the