account: Improve URI completion for local accounts
When providing account_uri_complete with an incomplete dialstring and
an account, it would always append the local URI of the account to the
dialstring - regardless of the account config. However, when using a
local account (with regint=0), this could cause BareSIP to try calling
itself.

Now, for an incomplete URI with account regint=0, the local host and
port are no longer appended to the URI. (Resulting in the supplied URI
being interpreted as hostport instead of userinfo.) One ua test was
modified to reflect this new behavior.

Further, a few potential memory leaks were fixed in modules using this
function.