call: allocate streams after peer_uri was set
For now the peer_uri should be available when the streams are allocated. In a
next PR it could be passed to the streams.

Maybe in another PR the streams could be allocated after the SDP negotiation.
Currently this is not possible due to the "common audio or video codecs" check
which depends on the audio/video streams.