call: hangup call on transp reset if necessary
If the transport is reset (e.g. an IP address change) and a call is in
the early media state, session modifications might not be possible. In
this case, the peer will never receive our transport update (e.g our
new IP) and we will no longer be able to communicate with the peer. In
such cases - when target refreshes are not possible - hangup the call.