variadic function fixes (#2523)
* ctrl_dbus,ice,png_vf: Fix format string usage

Invocations of functions with variable argument lists may never pass
string variables directly, but should always use "%s" as format string
and then the string as a variable.

* ua: fix append_params

If a parameter string contains characters like '[' (which are
specifically allowed according to RFC 3261) this could cause crashes
because the parameter string is passed to the re_regex function.

* sdp: fix sdp_decode_multipart

The multipart SDP boundary string is set externally and then passed to
the re_regex function. This could lead to crashes. Now, strstr() is
used.

* sdp & ua: Very minor improvements & formatting