call: never set sent_answer to false (#2674)
This flag is true if any SDP answer was already sent to the peer. It is used
in static_menu.c to decide if the `ldir` has to be set or not. After sending
the first SDP answer `ldir` should stay unchanged at the value that was used
during encoding the SDP answer. It is important for call_need_modify() to
decide if a re-INVITE is needed or not. That's why sent_answer should not be
reset to false. This could lead to a missing re-INVITE.