call: add Refer-To angle brackets
"The Contact, From, and To header fields contain a URI.  If the URI
contains a comma, question mark or semicolon, the URI MUST be
enclosed in angle brackets (< and >).  Any URI parameters are
contained within these brackets.  If the URI is not enclosed
in angle brackets, any semicolon-delimited parameters are
header-parameters, not URI parameters."