jbuf: wish size is given in #packets
The wish size for the adaptive mode depends on how much the packets are re-
ordered. It can only be measured by looking on the sequence number of the RTP
header and how much it jumps around because of re-ordered packets. Counting
video frames is not needed and causes failures when the wish size is computed.