menu: add follow up invite timer
This change fixes an issue that occurs by calling `ua_connect()` in the UA
event handler. The OUTGOING event was sent to the application before the
REFER or REDIRECT events that should trigger a follow up SIP call.

The UA events should be sent in this order to the application:
- REFER, then OUTGOING
- REDIRECT, then OUTGOING