portaudio: fixed selecting playback-only devices by name (refs #3330)
find_device() was looking only in the ausrc source device list.
If a playback device does not have input channels, it won't be in the
ausrc device list and cannot be found, which broke the `audio_player`
command for such devices.

We now pass the exact device list to find_device() as well.