call: do not go back to RINGING state if already in EARLY
In some cases, for example when 100rel is used, 1XX responses may arrive
out of order (e.g. due to re-transmissions with 100rel). If a 180
Ringing response arrives after a 183 response has already been
processed, it does not make much sense to stop the early media streams
and go back to ringing. So now, once in the EARLY state, the call will
not go back to the RINGING state.