menu: Add hangup sound on closed calls (#3425)
Gives the user the option to play an audio file when the call closes.
This is a common way in many phones to provide the user an additional
audio feedback when a call is terminated.

The new config option is called 'hangup_aufile' and is set to 'none' per
default. Meaning there is no playback on hangup unless the user
specifies a file.

Signed-off-by: Daniel Danzberger <dd@embedd.com>
Co-authored-by: Daniel Danzberger <dd@embedd.com>