gst: fix crash on shutdown
This fixes crash of gst_deinit() when gst module is loaded, not used and
baresip is shutdown.

The Doxygen of gst_deinit() tells that it is not needed to call this function.
It can lead to crashes. E.g. if gstreamer is used also in other modules.

It is no problem to call gst_init() multiple times. But calling
gst_deinit() more than once is not allowed. On the other hand, gstreamer does
its de-initialization automatically when the program terminates.