call: send local SDP event not too early (#3516)
* call: send local SDP event not too early

The BEVENT_CALL_LOCAL_SDP should not be sent before the SDP media
objects are available. Pre-requirement for SDP media is that the audio/video
streams exist already. The sipsess_desc_handler() is the last stage
where the streams have to be created.

* call: send LOCAL_SDP only once

Fix also error handling. The application may also decide to abort
sending the INVITE by processing the LOCAL_SDP event.